notifications

+++ social_media_api/posts/views.py
@@ -1,4 +1,4 @@
-from rest_framework import viewsets, permissions, filters, status
+from rest_framework import generics, viewsets, permissions, filters, status
 from .models import Post, Comment, Like
 from rest_framework.response import Response
 from .serializers import PostSerializer, CommentSerializer, LikeSerializer
@@ -49,13 +49,9 @@ class LikePostView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, pk):
-        try:
-            post = Post.objects.get(pk=pk)
-        except Post.DoesNotExist:
-            return Response({"error": "Post not found"}, status=status.HTTP_404_NOT_FOUND)
+        post = generics.get_object_or_404(Post, pk=pk)
 
         like, created = Like.objects.get_or_create(user=request.user, post=post)
-
         if not created:
             return Response({"message": "Already liked"}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -70,14 +66,11 @@ class UnlikePostView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, pk):
-        try:
-            post = Post.objects.get(pk=pk)
-        except Post.DoesNotExist:
-            return Response({"error": "Post not found"}, status=status.HTTP_404_NOT_FOUND)
+        post = generics.get_object_or_404(Post, pk=pk) 
 
-        try:
-            like = Like.objects.get(user=request.user, post=post)
+        like = Like.objects.filter(user=request.user, post=post).first()
+        if like:
             like.delete()
             return Response({"message": "Post unliked"}, status=status.HTTP_200_OK)
-        except Like.DoesNotExist:
-            return Response({"error": "You have not liked this post"}, status=status.HTTP_400_BAD_REQUEST)
+
+        return Response({"error": "You have not liked this post"}, status=status.HTTP_400_BAD_REQUEST)
